+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Renamed FormatterTextBlock to ChapterTextBlock
 - Renamed HeaderItemsControl to ChapterHeaderedContentControl
 - Renamed ImageButton to ChapterButton
+- Renamed ItemsPanel to ChapterItemsPanel
 ### Supported .Net Versions
 - .Net Core 3.0
 - .Net Framework 4.5